Fixing F0 logic

+++ src/sppysound/analysis/FFTAnalysis.py
@@ -85,7 +85,7 @@ class FFTAnalysis(Analysis):
         hopSize = int(frameSize - np.floor(overlapFac * frameSize))
 
         # zeros at beginning (thus center of 1st window should be for sample nr. 0)
-        samples = np.append(np.zeros(np.floor(frameSize/2.0)), sig)
+        samples = np.append(np.zeros(np.floor(frameSize/2).astype(int)), sig)
         # cols for windowing
 
         cols = np.ceil((len(samples) - frameSize) / float(hopSize)) + 1
